Use Array{Int} in similar() type assertions for 32-bit.
similar(::SLArray{Int}) stores a native Int vector; comparing against
Array{Int64} fails when Int===Int32.

+++ test/slarrays.jl
@@ -29,11 +29,11 @@ import ArrayInterface
     @test @inferred(similar_type(b, Float64, Size(3, 3))) ===
         SArray{Tuple{3, 3}, Float64, 2, 9}
 
-    @test typeof(@inferred(similar(b))) === LArray{Int, 1, Array{Int64, 1}, (:a, :b, :c)}
+    @test typeof(@inferred(similar(b))) === LArray{Int, 1, Array{Int, 1}, (:a, :b, :c)}
     @test typeof(@inferred(similar(b, Float64))) ===
         LArray{Float64, 1, Array{Float64, 1}, (:a, :b, :c)}
     @test typeof(@inferred(similar(b, Size(1, 3)))) ===
-        LArray{Int, 2, Array{Int64, 2}, (:a, :b, :c)}
+        LArray{Int, 2, Array{Int, 2}, (:a, :b, :c)}
     @test typeof(@inferred(similar(b, Size(3, 3)))) === MArray{Tuple{3, 3}, Int, 2, 9}
 
     @test @inferred(copy(b)) === ABC_int(Tuple(b))
